Use IVec2 for all pixel coordinates

+++ internal/window.go
@@ -24,8 +24,8 @@ import "C"
 var Window struct {
 	window  *C.SDL_Window
 	context C.SDL_GLContext
-	Width   float32
-	Height  float32
+	Width   int32
+	Height  int32
 }
 
 // Focus state
@@ -39,7 +39,7 @@ var (
 // OpenWindow creates the game window and its associated OpenGL context.
 func OpenWindow(
 	title string,
-	resolution [2]float32,
+	resolution [2]int32,
 	display int,
 	fullscreen bool,
 	fullscreenMode string,
